refactor deck of cards, elemental scroll, enchantment siphon, encouragement, fever break, flail, and flame scythes
