+++ source/gear/_gearDictionary.js
@@ -32,29 +32,13 @@ for (const file of [
 	"cursed-shield.js",
 	"cursed-tome.js",
 	"daggers.js",
-	"deckofcards-base.js",
-	"deckofcards-evasive.js",
-	"deckofcards-numbing.js",
-	"deckofcards-omenous.js",
-	"deckofcards-tormenting.js",
-	"elementalscroll-balanced.js",
-	"elementalscroll-base.js",
-	"elementalscroll-surpassing.js",
-	"enchantmentsiphon-base.js",
-	"enchantmentsiphon-flanking.js",
-	"enchantmentsiphon-tormenting.js",
-	"encouragement-base.js",
-	"encouragement-rallying.js",
-	"encouragement-vigorous.js",
-	"feverbreak-base.js",
-	"feverbreak-fatiguing.js",
-	"feverbreak-unstoppable.js",
-	"flail-base.js",
-	"flail-bouncing.js",
-	"flail-incompatible.js",
-	"flamescythes-base.js",
-	"flamescythes-thiefs.js",
-	"flamescythes-toxic.js",
+	"deck-of-cards.js",
+	"elemental-scroll.js",
+	"enchantment-siphon.js",
+	"encouragement.js",
+	"fever-break.js",
+	"flail.js",
+	"flame-scythes.js",
 	"flourish-base.js",
 	"flourish-bouncing.js",
 	"flourish-shattering.js",
